added movie list container to separate file
