+++ lib/domain/blocs/movie/now_playing_movie_list_bloc.dart
@@ -1,6 +1,7 @@
 import 'package:bloc_concurrency/bloc_concurrency.dart';
 import 'package:comics_db_app/configuration/configuration.dart';
 import 'package:comics_db_app/domain/api_client/movie_and_tv_api_client.dart';
+import 'package:comics_db_app/domain/blocs/movie/movie_list_container.dart';
 import 'package:comics_db_app/domain/blocs/movie/movie_popular_list_bloc.dart';
 import 'package:bloc/bloc.dart';
 import 'package:comics_db_app/domain/entity/movie.dart';
@@ -58,7 +59,7 @@ class NowPlayingMovieListBloc extends Bloc<MovieListEvent, MovieListState> {
   }
 
   Future<void> onNowPlayingMovieListEventLoadReset(MovieListEventLoadReset event, Emitter<MovieListState> emit) async {
-    emit(MovieListState.initial());
+    emit(const MovieListState.initial());
   }
 
   Future<void> onNowPlayingMovieListEventLoadSearch(MovieListEventSearchMovie event, Emitter<MovieListState> emit) async {
