+++ lib/domain/blocs/movie/top_rated_movie_list_bloc.dart
@@ -2,6 +2,7 @@ import 'package:bloc/bloc.dart';
 import 'package:bloc_concurrency/bloc_concurrency.dart';
 import 'package:comics_db_app/configuration/configuration.dart';
 import 'package:comics_db_app/domain/api_client/movie_and_tv_api_client.dart';
+import 'package:comics_db_app/domain/blocs/movie/movie_list_container.dart';
 import 'package:comics_db_app/domain/blocs/movie/movie_popular_list_bloc.dart';
 import 'package:comics_db_app/domain/entity/movie.dart';
 import 'package:comics_db_app/domain/entity/movie_response.dart';
@@ -60,7 +61,7 @@ class TopRatedMovieListBloc extends Bloc<MovieListEvent, MovieListState> {
   }
 
   Future<void> onTopRatedMovieListEventLoadReset(MovieListEventLoadReset event, Emitter<MovieListState> emit) async {
-    emit(MovieListState.initial());
+    emit(const MovieListState.initial());
   }
 
   Future<void> onTopRatedMovieListEventLoadSearch(MovieListEventSearchMovie event, Emitter<MovieListState> emit) async {
